added: book images + login + create/delete etc

+++ controllers/delete.php
@@ -2,10 +2,13 @@
 
 switch ($_SERVER['REQUEST_METHOD']) {
     case 'POST':
-        $config = require "config.php";
-        require "Database.php";
-        $db = new Database($config);
-        $db->execute("DELETE FROM books WHERE id=:id", [":id" => $_POST['id']]);
+        session_start();
+        if ($_SESSION['user']['permission_level'] ?? 0 > 0) {
+            $config = require "config.php";
+            require "Database.php";
+            $db = new Database($config);
+            $db->execute("DELETE FROM books WHERE id=:id", [":id" => $_POST['id']]);
+        }
         header("Location: /books");
         break;
     default:
